bug fixes in controllers

+++ server/src/utils/cloudnary.js
@@ -21,7 +21,9 @@ const uploadOnCloudinary = async (localFilePath) => {
     fs.unlinkSync(localFilePath);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath);
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
     //remove the locally saved temp file as upload got failed
     return null;
   }
